LayoutTests/fast/forms/month/month-editable-components/month-editable-components-mouse-events-expected.txt

Test for mouse events for <input type=month>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Enabled Input

PASS input.value is "2020-09"
PASS input.value is "0012-09"
PASS input.value is "0012-05"
PASS input.value is "0012-05"
PASS clickEventsFired is 3

Disabled Input

PASS clickEventsFired is 0

Readonly Input

PASS clickEventsFired is 3
PASS successfullyParsed is true

TEST COMPLETE
